--- packages/geo/src/hexworld/index.ts
@@ -1,9 +1,0 @@
-export * from "./bands.ts";
-export * from "./borders.ts";
-export * from "./cells.ts";
-export * from "./countries.ts";
-export * from "./detail-zone.ts";
-export * from "./land.ts";
-export * from "./style.ts";
-export * from "./wedges.ts";
-export * from "../shared/index.ts";
