+++ packages/geo/src/index.ts
@@ -1,4 +1,6 @@
-export * from "./components/ak-map.ts";
-export * from "./style.ts";
-export * from "./hexworld/index.ts";
-export * from "./shared/index.ts";
+export * from "./borders.ts";
+export * from "./cells.ts";
+export * from "./countries.ts";
+export * from "./detail-zone.ts";
+export * from "./labels.ts";
+export * from "./land.ts";
