--- packages/geo/src/shared/events.ts
@@ -1,3 +1,0 @@
-import { EventActions } from "@goauthentik/api/src/models/EventActions.ts";
-
-export type EventActionsRecord = { [key in EventActions]?: string };
