--- packages/geo/src/shared/index.ts
@@ -1,2 +1,0 @@
-export * from "./events.ts";
-export * from "./labels.ts";
