--- packages/geo/src/types.d.ts
@@ -1,6 +1,0 @@
-declare module "*.css" {
-    import { CSSResult } from "lit";
-
-    const css: CSSResult & { readonly __brand?: string };
-    export default css;
-}
