+++ web/src/elements/maps/events.ts
@@ -1,0 +1,3 @@
+import { EventActions } from "@goauthentik/api";
+
+export type EventActionsRecord = { [key in EventActions]?: string };
